% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\name{evalidators}
\alias{evalidators}
\title{Evaluate Validators safely}
\usage{
evalidators(L)
}
\arguments{
\item{L}{A list resulting from combining several validators}
}
\description{
Evaluate Validators safely
}
