% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\name{inspector}
\alias{inspector}
\title{data inspector}
\usage{
inspector(x)
}
\arguments{
\item{x}{a data.table with its (S3) class extended by the database table name (see server.R)}
}
\description{
Data police
}
\details{
Inspectors are usually a collection of validators
}
\examples{
 require(data.table)
 require(magrittr)
 x = data.table(
     v1        = c(NA, NA, as.character(Sys.time() - 3600*24*10 )  ) ,
     datetime_ = c('2016-11-23 25:23', as.character(Sys.time() -100) ,as.character(Sys.time()+100))
     )
 class(x) = c(class(x), 'tablex')

 inspector.tablex <- function(x) {
  list(
      x[, .(datetime_)]  \%>\% POSIXct_validator, 
      x  \%>\% is.na_validator
   )
  }

 evalidators( inspector(x) )

}
