% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\name{inspector_loader}
\alias{inspector_loader}
\title{Inspector loader}
\usage{
inspector_loader(package)
}
\arguments{
\item{package}{a package name}
}
\description{
Load inspectors of a given package
}
\examples{

inspector_loader(package = 'DataEntry')

}
