% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\name{try_validator}
\alias{try_validator}
\title{Let a validator fail without an error}
\usage{
try_validator(..., nam = "")
}
\arguments{
\item{nam}{an optional name to help identify the failty validator later.}

\item{x}{an expression that should tipically return a validator result but in some,
yet to be discovered, exceptional cases, it does not.}
}
\description{
Let a validator fail without an error
}
\examples{
x = data.table(id = 1:10)
wrong_output_validator <- nrow
faulty_validator <- function(x) foo(x)

faulty_validator (x) |> try_validator(nam = 'foo')
nrow(x) |> try_validator()

}
