% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\name{inspector_loader}
\alias{inspector_loader}
\title{Inspector loader}
\usage{
inspector_loader(path)
}
\arguments{
\item{path}{The path to the files containing the inspector functions.}
}
\description{
Loads inspectors from a file.
}
\note{
This function is just sourcing the file but it does not return an error when something goes wrong.
}
\examples{
p <- system.file("inspectors", "DataEntry.R", package = "DataEntry.validation")
inspector_loader(p)

}
